Move ranking helpers to it own ranking.py file.
