+++ tests/tournaments/test_helpers.py
@@ -5,7 +5,7 @@ from datetime import date
 
 from django.test import TestCase
 
-from tournaments.helpers import compute_ranking_positions
+from tournaments.ranking import compute_ranking_positions
 from tournaments.models import PadelRanking, Person
