+++ tournaments/management/commands/misc.py
@@ -1,6 +1,6 @@
 from django.core.management.base import BaseCommand
 
-from tournaments.helpers import compute_ranking_positions, compute_ranking_tournaments
+from tournaments.ranking import compute_ranking_positions, compute_ranking_tournaments
 
 
 class Command(BaseCommand):
